\name{fn_parameter_mleEM}
\alias{fn_parameter_mleEM}
\title{fn_parameter_mleEM}
\usage{
fn_parameter_mleEM(ls_data,v_r_observedTerm)
}
\description{
Calculates the maximum likelihood estimates of the payment and settlement rates.
}
\arguments{
\item{ls_data}{A dataset in the form of a list}
\item{v_r_observedTerm}{A vector of reals representing the observed term of each claim}
}

