\name{fn_pseudoData}
\alias{fn_pseudoData}
\title{fn_pseudoData}
\usage{
fn_pseudoData(n_numClaims,v_r_observedTerm,v_r_paymentRate,v_r_settleRate)
}
\description{
Generates a pseudo dataset in the form of a list for use in the package's functions.
}
\arguments{
\item{n_numClaims}{The number of claims that should be included in the dataset}
\item{v_r_observedTerm}{A vector of reals representing the observed term of each claim}
\item{v_r_paymentRate}{The payment rate of each claim}
\item{v_r_settleRate}{The settlement rate of each claim}
}
